fixed c070ff2 from https://gitee.com/liu-zhiyuan1995/graphic_graphic_2d_RDPB/pulls/2400 correct YCodeCheck

+++ rosen/modules/2d_graphics/src/render_context/cache_data.h
@@ -61,7 +61,7 @@ private:
     bool IfSkipClean(const size_t addedSize);
     bool IfCleanFinished();
     void RandClean(const size_t cleanThreshold);
-    size_t Clean(const int removeIndex);
+    size_t Clean(const size_t removeIndex);
 
     static inline size_t Align4(size_t size)
     {
